Finish factoring out memory allocation
